feat(smallProductInfo): 작은 제품 정보 컴포넌트 작성.
