+++ src/client/src/components/organisms/Floater/index.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent, useState } from "react";
-import Image from "next/image";
 
 import Button from "components/atoms/Button";
 import Icon from "components/atoms/Icon";
 import Knob from "components/atoms/Knob";
+import ProductSmallInfo from "components/molecules/ProductSmallInfo";
 
 import styled from "@emotion/styled";
 
@@ -42,15 +42,12 @@ const Floater: FunctionComponent<FloaterProps> = (props) => {
 	return (
 		<FloaterWrapper>
 			<InnerBox>
-				<ProductArea>
-					<ProductThumb backgroundColor={backgroundColor}>
-						<StyledImg src={imgSrc} />
-					</ProductThumb>
-					<ProductInfoArea>
-						<StyledProductName>{productName}</StyledProductName>
-						<StyledProductNameKor>{productNameKor}</StyledProductNameKor>
-					</ProductInfoArea>
-				</ProductArea>
+				<ProductSmallInfo
+					imgSrc={imgSrc}
+					backgroundColor={backgroundColor}
+					productName={productName}
+					productNameKor={productNameKor}
+				/>
 				<ButtonArea>
 					<Button
 						style={{
@@ -105,51 +102,6 @@ const InnerBox = styled.div`
 	display: flex;
 	align-items: center;
 	justify-content: space-between;
-`;
-
-const ProductArea = styled.div`
-	padding-right: 40px;
-	display: flex;
-`;
-
-const ProductThumb = styled.div<{ backgroundColor: string }>`
-	background-color: ${({ backgroundColor }) => backgroundColor};
-	flex: none;
-	margin-right: 12px;
-	width: 64px;
-	height: 64px;
-	border-radius: 12px;
-`;
-
-const StyledImg = styled(Image)`
-	width: 100%;
-	vertical-align: top;
-`;
-
-const ProductInfoArea = styled.div`
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	flex: 1;
-`;
-
-const StyledProductName = styled.p`
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-	line-height: 18px;
-	font-size: 15px;
-	margin: 10px 0 0 0;
-`;
-
-const StyledProductNameKor = styled.p`
-	line-height: 14px;
-	font-size: 12px;
-	letter-spacing: -0.06px;
-	color: rgba(34, 34, 34, 0.5);
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
 `;
 
 const ButtonArea = styled.div`
