+++ src/client/src/components/organisms/HeaderMain/index.tsx
@@ -1,4 +1,4 @@
-import React, { FunctionComponent, useState } from "react";
+import React, { FunctionComponent } from "react";
 import Link from "next/link";
 
 import Logo from "components/atoms/Logo";
